back/requirements.txt
Flask
Firebase
Firebase-admin
pyrebase4